Extract web chat UI event mapping
